MFC r302784:
Fix a page fault in ip6_setpktopt(), occurring when the pflog module is
loaded, and syncthing is started, which uses setsockopt(IPV6_PKGINFO).

This is because pflog interfaces do not normally have an IPv6 address,
causing the ND_IFINFO() macro to dereference a NULL pointer.

Approved by:	re (kib)
Reviewed by:	ae
PR:		210943


git-svn-id: https://svn.freebsd.org/base/stable/11@302934 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f