Add my beinstall script.
This is meant to install a new BE (boot environment) given a fully built
world/kernel.  In addition to installing world and kernel in the new BE,
it also automatically performs /etc updates (using etcupdate or mergemaster)
and package updates (using pkg).

Because this process is performed in a new BE, it reduces the need for a
second reboot.  It also means a reboot into a partially updated system (due
to install or hardware failure) can't happen.

Inspired by and similar in function to Solaris/illumos-style upgrades.


git-svn-id: https://svn.freebsd.org/base/head@302977 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f