MFC r302278:   libcxxrt: correct mangled "typeinfo name" symbols in Version.map
r261644 (MFC of r260553) added missing C++ typinfos to libcxxrt's
version script. It appears that a number of duplicate mangled symbols
were added due to a cut and paste error. Switch the second instances to
_ZTS*, typeinfo name for *.

Found by lld, which produces an error or warning for duplicate symbols.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/10@303293 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f