MFC r303034: Include makewhatis in ITOOLS when MK_MAN_UTILS is true
Previously it was conditional on MK_MAN. It's possible to build
FreeBSD with man pages but without man page tools. MK_MAN_UTILS
is the conditional used in share/man/Makefile for determining whether
makewhatis is executed at install time, so it is the proper one for
ITOOLS as well.

PR:		210142
Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/stable/11@303337 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f