MFC 303109: Update crashinfo to work with newer gdb from ports.
If gdb from ports is installed, use it instead of the base system gdb
to extract variables from a kernel.  Note that base gdb and ports gdb
do not support the same options for invoking a single command in batch
mode, so a wrapper shell function is used.  In addition, prefer kgdb
from ports when generating a backtrace if present.

PR:		193335


git-svn-id: https://svn.freebsd.org/base/stable/9@303403 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f