MFC r303788
Fix some logic in PCIe HotPlug; display EI status

The interpretation of the Electromechanical Interlock Status was
inverted, so we disengaged the EI if a card was inserted.
Fix it to engage the EI if a card is inserted.

When displaying the slot capabilites/status with pciconf:

- We inverted the sense of the Power Controller Control bit,
  saying the power was off when it was really on (according to
  this bit).  Fix that.

- Display the status of the Electromechanical Interlock:
	EI(engaged)
	EI(disengaged)

Approved by:	re (gjb)
Sponsored by:	Dell Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@303835 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f